HistoricalPrice Endpoint and Tests

+++ src/main/java/org/galatea/starter/domain/IexHistoricalPrice.java
@@ -10,10 +10,10 @@ public class IexHistoricalPrice {
 
   private String symbol;
   private String date;
-  private Integer volume;
+  private long volume;
   private BigDecimal close;
   private BigDecimal high;
   private BigDecimal low;
   private BigDecimal open;
-  
+
 }
